fix package and add sourceSets in build.gradle
