--- src/main/java/Main.java
@@ -1,12 +1,0 @@
-package src.main.java;
-
-import src.main.java.cortana.Cortana;
-
-public class Main {
-    
-    public static void main(String[] args) {
-        Cortana cortana = new Cortana();
-        cortana.run();
-    }
-
-}
