+++ src/main/java/chatbot/cortana/Parser.java
@@ -1,14 +1,14 @@
-package src.main.java.cortana;
+package chatbot.cortana;
 
 import java.time.LocalDateTime;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import src.main.java.exception.InvalidInputException;
-import src.main.java.task.DeadlineTask;
-import src.main.java.task.EventTask;
-import src.main.java.task.TaskList;
-import src.main.java.task.TodoTask;
+import chatbot.exception.InvalidInputException;
+import chatbot.task.DeadlineTask;
+import chatbot.task.EventTask;
+import chatbot.task.TaskList;
+import chatbot.task.TodoTask;
 
 public class Parser {
