+++ src/main/java/chatbot/task/DeadlineTask.java
@@ -1,4 +1,4 @@
-package src.main.java.task;
+package chatbot.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
