+++ src/main/java/chatbot/task/TodoTask.java
@@ -1,4 +1,4 @@
-package src.main.java.task;
+package chatbot.task;
 
 public class TodoTask extends Task {
